refactor: 몽고 쿼리 수 최적화

+++ src/main/java/com/github/kmu_wink/seoul_in_culture/domain/chat/dto/response/RoomListResponse.java
@@ -17,7 +17,7 @@ public record RoomListResponse(
 
             Meeting meeting,
             Chat last,
-            int unread
+            long unread
     ) {
 
     }
