+++ src/main/java/com/github/kmu_wink/seoul_in_culture/domain/user/dto/response/GetMyInfoResponse.java
@@ -15,7 +15,7 @@ public record GetMyInfoResponse(
 
         Collection<Event> bookmarks,
 
-        int joinedMeetings,
+        long joinedMeetings,
         Collection<Meeting> hostedMeetings,
 
         Collection<Review> reviews,
