+++ src/main/java/com/github/kmu_wink/seoul_in_culture/domain/user/dto/response/GetOtherInfoResponse.java
@@ -12,9 +12,9 @@ public record GetOtherInfoResponse(
 
         User user,
 
-        int bookmarks,
+        long bookmarks,
 
-        int joinedMeetings,
+        long joinedMeetings,
         Collection<Meeting> hostedMeetings,
 
         Collection<Review> reviews,
